requirements.txt
# Dépendances Python pour SDAU Zorgho
Django==4.2.7
psycopg2-binary==2.9.9
# GeoDjango et cartographie

# Alternative si GDAL pose problème : utiliser les binaires pré-compilés
# Pour Windows : https://www.lfd.uci.edu/~gohlke/pythonlibs/#gdal

# Extensions géospatiales
geopandas==0.14.1
shapely==2.0.2

# API REST
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-cors-headers==4.3.1

# Authentification JWT
djangorestframework-simplejwt==5.3.0

# Filtrage
django-filter==23.5

# Variables d'environnement
python-decouple==3.8

# Sécurité
django-environ==0.11.2

# Utilitaires
pillow==10.1.0
python-dateutil==2.8.2

# Développement
django-extensions==3.2.3
